test: align process supervisor ownership
